requirements_gpu.txt
# GPU优化版依赖文件 - 针对RTX 4090 Linux环境
# 安装命令: pip install -r requirements_gpu.txt

# 核心框架
langchain>=0.1.0
langgraph>=0.0.40
langchain-community>=0.0.20
langchain-core>=0.1.0

# LLM集成
langchain-ollama>=0.1.0

# 向量数据库和嵌入（GPU优化版本）
chromadb>=0.4.0
sentence-transformers>=2.2.0

# PyTorch GPU版本（CUDA 11.8支持）
--index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0+cu118
torchvision>=0.15.0+cu118
torchaudio>=2.0.0+cu118

# Transformers和加速库
transformers>=4.30.0
accelerate>=0.20.0

# GPU加速向量搜索
faiss-gpu>=1.7.4

# CUDA支持库
cupy-cuda12x>=12.0.0

# 文档处理
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# 网络搜索
tavily-python>=0.3.0

# 数据处理（兼容版本）
numpy>=1.24.0,<2.0
pandas>=2.0.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# 系统监控
gputil>=1.4.0
psutil>=5.9.0

# 可选：Jupyter支持
jupyter>=1.0.0
ipykernel>=6.0.0